transformer/requirements.txt
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
idna==3.4
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.0.32
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
protobuf==3.20.3
psycopg2-binary==2.9.6
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
